+++ docs/flutter_service_worker.js
@@ -23,14 +23,14 @@ const RESOURCES = {"assets/AssetManifest.bin": "693635b5258fe5f1cda720cf224f158c
 "canvaskit/skwasm.worker.js": "bfb704a6c714a75da9ef320991e88b03",
 "favicon.png": "5dcef449791fa27946b3d35ad8803796",
 "flutter.js": "383e55f7f3cce5be08fcf1f3881f585c",
-"flutter_bootstrap.js": "9c40d47375513913db0cd184d20a71a2",
+"flutter_bootstrap.js": "faab2580e2a8dcf47fef0299dbe73893",
 "icons/Icon-192.png": "ac9a721a12bbc803b44f645561ecb1e1",
 "icons/Icon-512.png": "96e752610906ba2a93c65f8abe1645f1",
 "icons/Icon-maskable-192.png": "c457ef57daa1d16f64b27b786ec2ea3c",
 "icons/Icon-maskable-512.png": "301a7604d45b3e739efc881eb04896ea",
-"index.html": "f95a9a5d93937164d848c062fa2cddd2",
-"/": "f95a9a5d93937164d848c062fa2cddd2",
-"main.dart.js": "277f847ddaaadcbcb1154551ebe7f038",
+"index.html": "198aafef5b85ce5fe29056ba6024bf1f",
+"/": "198aafef5b85ce5fe29056ba6024bf1f",
+"main.dart.js": "fa9db3461ae252ed056986545db11fc3",
 "manifest.json": "9a0de1ff26661cd281cb9deb3db33b02",
 "version.json": "ff966ab969ba381b900e61629bfb9789"};
 // The application shell files that are downloaded before a service worker can