+++ lib/flutter_graph_widget.dart
@@ -1,6 +1,7 @@
 // Copyright (c) 2023- All flutter_graph_view authors. All rights reserved.
 //
 // This source code is licensed under Apache 2.0 License.
+import 'package:flame/extensions.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_graph_view/flutter_graph_view.dart';
@@ -13,6 +14,8 @@ class FlutterGraphWidget extends StatefulWidget {
   final GraphAlgorithm algorithm;
   final DataConvertor convertor;
   final Options? options;
+
+  static bool stopDrag = false;
 
   const FlutterGraphWidget({
     Key? key,
@@ -36,22 +39,61 @@ class _FlutterGraphWidgetState extends State<FlutterGraphWidget> {
     super.initState();
   }
 
+  @override
+  void dispose() {
+    super.dispose();
+    graphCpn.removePauseListener();
+  }
+
   late GraphComponent graphCpn;
   Map<String, Widget Function(BuildContext, GraphComponent)>
       overlayBuilderMap2 = {};
 
   @override
   Widget build(BuildContext context) {
-    return GameWidget(
-      backgroundBuilder: widget.options?.backgroundBuilder,
-      overlayBuilderMap: overlayBuilderMap2,
-      game: graphCpn = GraphComponent(
-        data: widget.data,
-        convertor: widget.convertor,
-        algorithm: widget.algorithm,
-        options: widget.options,
-        context: context,
-      ),
+    double? scale;
+    Vector2 center = Vector2.zero();
+    return Stack(
+      children: [
+        Positioned.fill(
+          child: GameWidget(
+            backgroundBuilder: widget.options?.backgroundBuilder,
+            overlayBuilderMap: overlayBuilderMap2,
+            game: graphCpn = GraphComponent(
+              data: widget.data,
+              convertor: widget.convertor,
+              algorithm: widget.algorithm,
+              options: widget.options,
+              context: context,
+            ),
+          ),
+        ),
+        Positioned.fill(
+            child: MouseRegion(
+          onHover: (ev) => center = ev.localPosition.toVector2(),
+          opaque: false,
+        )),
+        Positioned.fill(
+            child: GestureDetector(
+          onScaleUpdate: (ScaleUpdateDetails details) {
+            if (FlutterGraphWidget.stopDrag) return;
+            // single finger pan
+            if (details.pointerCount == 1) {
+              var v = details.focalPointDelta.toVector2();
+              graphCpn.onPanUpdate(v);
+            } else if (details.pointerCount == 2) {
+              graphCpn.onZoom(
+                zoomCenter: center,
+                zoomDelta: scale == null
+                    ? (details.scale - 1)
+                    : (details.scale - scale!),
+              );
+              scale = details.scale;
+            }
+          },
+          onScaleEnd: (details) => scale = null,
+        ))
+      ],
     );
   }
 }