+++ lib/model/edge.dart
@@ -77,25 +77,34 @@ class Edge {
   double get computeIndex {
     var edgeList = cpn?.graph.edgesFromTwoVertex(start, end) ?? [];
     var idx = edgeList.indexOf(this);
+    var result = 0.0;
     if (edgeList.length.isOdd) {
       if (idx.isEven) {
-        return idx / 2;
+        result = idx / 2;
       } else {
-        return -(idx + 1) / 2;
+        result = -(idx + 1) / 2;
       }
     } else {
       if (idx.isEven) {
-        return idx / 2 + 0.5;
+        result = idx / 2 + 0.5;
       } else {
-        return -(idx - 1) / 2 - 0.5;
+        result = -(idx - 1) / 2 - 0.5;
       }
     }
+
+    return start.id.hashCode > end!.id.hashCode ? -result : result = result;
   }
 
   int get edgeIdx {
     var edgeList = cpn?.graph.edgesFromTwoVertex(start, end) ?? [];
     var idx = edgeList.indexOf(this);
     return idx;
+  }
+
+  double get edgeIdxRatio {
+    var edgeList = cpn?.graph.edgesFromTwoVertex(start, end) ?? [];
+    var ratio = edgeList.indexOf(this) / edgeList.length;
+    return ratio;
   }
 
   /// Uniqueness based on primary key of triple.
@@ -110,5 +119,6 @@ class Edge {
       (other is Edge &&
           other.ranking == ranking &&
           other.start == start &&
-          other.end == end);
+          other.end == end &&
+          other.edgeName == edgeName);
 }