+++ CHANGELOG.md
@@ -1,5 +1,132 @@
+## 1.2.0
+- feat: add `stopDrag` to `FlutterGraphWidget` to control the drag behavior of the graph view.
+    > 支持通过 `FlutterGraphWidget.stopDrag = false;` 停止视窗拖拽。
+- fix[#27]: the issue does not work properly in version 1.29.0.
+    > 在 1.29.0 版本中无法正常工作的问题 
+    
+### Behavior change:
+- Upgrade flame to 1.29.0, so the `gameRef` in `VertexComponent` should be replaced with `game`, when using `vertex.cpn?.gameRef`.
+    > 升级 flame 到 1.29.0 版本，所以 `VertexComponent` 中的 `gameRef` 应该替换为 `game`，用到 `vertex.cpn?.gameRef`时。
+
+## 1.1.6
+- feat: 为 edge 新增 `SolidArrowEdgeDecorator` 实心箭头装饰器
+    > support `SolidArrowEdgeDecorator` for edge
+
+## 1.1.5
+
+- feat: Support controlling the game pause or play through the `options.pause`.
+    > 支持通过 `options.pause` 控制游戏暂停或播放。
+- feat: Support specifying node components through `options.vertexComponentNew`
+    > 支持通过 `options.vertexComponentNew` 指定节点组件。
+
+## 1.1.4+1
+- fix: the issue of edge overlap when the edgeName is different but same src, dst and edgeId.
+
+## 1.1.4
+- feat: add interface to `GraphAlgorithm`: `afterDrag`, `beforeMerge` and etc.
+- feat: add `LegendDecorator` to support the legend of the graph, to control the vertex and edge display. And fix the scroll issue of the legend panel.
+    - feat: 新增 `LegendDecorator`，支持图的图例，控制节点和边的显示。并修复图例面板的滚动问题。
+    example:
+    ```dart
+    // First add LegendDecorator to your decorators, 
+    // then set the legend options.
+    options
+        ..useLegend = false
+    ```
+- behavior changed: remove `speed` property from `VertexComponent`.
+    > What will be effective is that all the vertex.position will be set to the vertex.cpn.position directly, when decorating the vertex in `GraphAlgorithm` and sub-classes.
+
+- behavior changed: set the radius of vertex to a private variable
+- feat: adding `GraphRouteDecorator`, make the data like a brower history.
+    - feat: 新增 `GraphRouteDecorator`，使数据像浏览器历史一样，支持前进后退，当有合并行为发生时。
+
+- feat: adding `PauseDecorator`, make the graph stop updating.
+    - feat: 新增 `PauseDecorator`，使图的节点停止位置更新。
+
+- feat: adding control panels of decorators.
+    example:
+    ```dart
+      /// 指定装饰器参数的控制面板创建方法
+      CoulombReverseDecorator(
+        handleOverlay: kCoulombReserseOverlayBuilder(),
+      ),
+    ```
+- feat: enable configuring the force factor between brother vertexs.
+    example:
+    ```dart
+    /// @en: Make the repulsion between similar points smaller
+    /// 
+    /// @zh: 使相似点之间的排斥力变小
+    CoulombReverseDecorator(sameSrcAndDstFactor: 1.1), 
+    ```
+- fix: the line added later covers the previous point.
+- feat: support image vertex.
+- feat: using tag similarity as a layout element.
+    example:
+    ```dart
+    /// @en: Make the repulsion between similar tags smaller
+    /// 
+    /// @zh: 使相似标签之间的排斥力变小
+    CoulombReverseDecorator(sameTagsFactor: 0.8), 
+    ```
+
+- feat: add `dragged` property to `VertexComponent` to indicate whether the vertex is being dragged.
+
+## 1.1.3+1
+- fix: the crash issue when force is not a number [ForceMotionDecorator].
+
+## 1.1.3
+- feat: `EdgeLineShape` can append decorators.
+    - feat: 使 `EdgeLineShape` 可以追加装饰器。
+    ```dart
+        Options()
+            ..edgeShape = EdgeLineShape(
+                decorators: [
+                    DefaultEdgeDecorator(),
+                ],
+            )
+    ```
+
+- fix: the edge overlap problem of multiple edge graphs with two identical nodes.
+- fix: Edge with the same parameters except for different edge type is misdiagnosed as the same edge.
+- fix: correct the position of the two points of the edge.
+    - fix: 修正边的两个点的位置。
+- fix: the overlap issue of the data panel.
+
+## 1.1.2
+- feat: support specifying legends and legend text build.
+    
+    example:
+    ```dart
+    Options()
+        ..legendBuilder = (color, i) { // default
+            return RectangleComponent.fromRect(
+                Rect.fromLTWH(40, 50.0 + 30 * i, 30, 18),
+                paint: Paint()..color = color,
+            );
+        }
+        ..legendTextBuilder = (tag, i, color, position) { // default
+            return TextComponent(
+                text: tag,
+                position: Vector2(position.x + 40, position.y - 6),
+                // position: Vector2(position.x + 40, position.y - 2),
+                // textRenderer: TextPaint(
+                //   style: TextStyle(
+                //     fontSize: 17.0,
+                //     color: Colors.white,
+                //   ),
+                // ),
+            );
+        }
+    ```
+
+## 1.1.1
+- fix usage of both Scale and Pan on mobile platforms. ( [#12](https://github.com/graph-cn/flutter_graph_view/pull/12), via: [Mykyta Sadchenko](https://github.com/muknta))
+- feat: support zooming through gestures.
+
 ## 1.1.0
 - feat: add interface to `GraphComponent`: addVertex, addEdge, mergeGraph
+- feat: add implementation of PersistenceDecorator to store position of vertex. ([#10](https://github.com/graph-cn/flutter_graph_view/pull/10) [#11](https://github.com/graph-cn/flutter_graph_view/pull/11), via: [jersonal-com](https://github.com/jersonal-com))
 
 ### Behavior change:
 - interface change: add a graph parameter to `DataConvertor.convertGraph`